+++ tests/Instances/Data/Bifunctor.hs
@@ -18,6 +18,7 @@ module Instances.Data.Bifunctor () where
 
 import Data.Bifunctor.Biff (Biff(..))
 import Data.Bifunctor.Clown (Clown(..))
+import Data.Bifunctor.Fix (Fix(..))
 import Data.Bifunctor.Flip (Flip(..))
 import Data.Bifunctor.Join (Join(..))
 import Data.Bifunctor.Joker (Joker(..))
@@ -32,6 +33,7 @@ import Test.QuickCheck (Arbitrary(..))
 
 deriving instance Arbitrary (p (f a) (g b)) => Arbitrary (Biff p f g a b)
 deriving instance Arbitrary (f a)           => Arbitrary (Clown f a b)
+deriving instance Arbitrary (p (Fix p a) a) => Arbitrary (Fix p a)
 deriving instance Arbitrary (p b a)         => Arbitrary (Flip p a b)
 deriving instance Arbitrary (p a a)         => Arbitrary (Join p a)
 deriving instance Arbitrary (g b)           => Arbitrary (Joker g a b)